# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lifecycle Policy Engine",
    "version": "1.0.0"
  },
  "paths": {
...
    "/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "List Assets",
        "operationId": "list_assets_assets_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Type"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "name": "vendor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search hostname, vendor, or model",
              "title": "Q"
            },
            "description": "Search hostname, vendor, or model"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetRead"
                  },
                  "title": "Response List Assets Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assets/{asset_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Asset",
        "operationId": "get_asset_assets__asset_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ingest/runzero": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Trigger Runzero Ingestion",
        "operationId": "trigger_runzero_ingestion_ingest_runzero_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/ingest/file": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest File",
        "operationId": "ingest_file_ingest_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_ingest_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/ingest/runs": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get Ingestion Runs",
        "operationId": "get_ingestion_runs_ingest_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngestionRunRead"
                  },
                  "type": "array",
                  "title": "Response Get Ingestion Runs Ingest Runs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AssetDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "asset_type": {
            "type": "string",
            "title": "Asset Type"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "environment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "identifiers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Identifiers"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "os_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "N",
                  "N-1",
                  "N-2",
                  "Unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Os Tier"
          },
          "firmware_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "N",
                  "N-1",
                  "N-2",
                  "Unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Tier"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationRead"
            },
            "type": "array",
            "title": "Observations"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRead"
            },
            "type": "array",
            "title": "Evaluations"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_type",
          "identifiers",
          "first_seen_at",
          "last_seen_at",
          "observations",
          "evaluations"
        ],
        "title": "AssetDetail"
      },
      "AssetRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "asset_type": {
            "type": "string",
            "title": "Asset Type"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "environment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "identifiers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Identifiers"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "os_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "N",
                  "N-1",
                  "N-2",
                  "Unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Os Tier"
          },
          "firmware_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "N",
                  "N-1",
                  "N-2",
                  "Unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Tier"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_type",
          "identifiers",
          "first_seen_at",
          "last_seen_at"
        ],
        "title": "AssetRead"
      },
...
      "Body_ingest_file_ingest_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_file_ingest_file_post"
      },
...
      "EvaluationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "component_type_id": {
            "type": "string",
            "title": "Component Type Id"
          },
          "component_type_key": {
            "type": "string",
            "title": "Component Type Key"
          },
          "observed_version": {
            "type": "string",
            "title": "Observed Version"
          },
          "lifecycle_tier": {
            "type": "string",
            "enum": [
              "N",
              "N-1",
              "N-2",
              "Unsupported"
            ],
            "title": "Lifecycle Tier"
          },
          "matched_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Policy Id"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          },
          "rationale": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rationale"
          },
          "asset_hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Hostname"
          },
          "asset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type"
          },
          "environment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "component_type_id",
          "component_type_key",
          "observed_version",
          "lifecycle_tier",
          "evaluated_at",
          "rationale"
        ],
        "title": "EvaluationRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestRunRequest": {
        "properties": {
          "search": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search"
          },
          "fields": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "title": "IngestRunRequest"
      },
      "IngestionRunRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at",
          "status",
          "source",
          "stats"
        ],
        "title": "IngestionRunRead"
      },
      "ObservationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "component_type_id": {
            "type": "string",
            "title": "Component Type Id"
          },
          "component_type_key": {
            "type": "string",
            "title": "Component Type Key"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "id",
          "component_type_id",
          "component_type_key",
          "version",
          "observed_at",
          "source",
          "raw"
        ],
        "title": "ObservationRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}